Refactor absen logic
1. Introduce absen_from_line wrapper/handler and make absen() not imported.
2. Move input sanity check to absen_from_line
3. Master now only gives absen_from_line the received text minus absensi key character identifier.
4. Absensi key character identifier changed from / to !
5. Move timeit to more thoroughly measure execution time from even loading up the webdriver.

This commit happens as a result of me trying to make absen() something like a one-off use disposable class, just to fit the rest of Master games membership relationship thingy. I realize the truth is absen() is NOT an instance that requires remembering, it's more like stateless request to the bot to 'load up Selenium and do this sequence and let me know what is the result'. I end up thinking that Master should not handle the input sanity check so I moved it to absen.py and add a wrapper/interface/whateveritiscalledincomputerscience absen_from_line()