Debug some errors
It turned out there is a misnamed variable kode_presensi that should be
kode_absen. The NameError is swallowed by the large try-except block
(and not printed).

To prevent such exception-swallowing to happen again, a print statement
for debug (can be seen in Heroku's logs) is added to the large
try-except block.